feat: drop image
