+++ package.json
@@ -35,7 +35,6 @@
     "next-translate": "^1.4.0",
     "react": "18.0.0",
     "react-dom": "18.0.0",
-    "react-dropzone": "^14.2.1",
     "react-hook-form": "^7.31.1",
     "react-json-view": "^1.21.3",
     "react-query": "^3.39.0",
